rag-gemini/requirements.txt
numpy
pandas
tqdm
google-genai
Pillow
ratelimit
PyMuPDF
python-dotenv
streamlit
